+++ src/python/exomiser-thresholding.py
@@ -204,8 +204,6 @@ def worker(queue, output_location):
             log_files = [file for file in files if file.endswith(".log")]
             vcf_files = [file for file in files if file.endswith(".vcf.gz")]
             if all(os.path.exists(log_file) for log_file in log_files):
-                print("Start monitor logs")
-                print(vcf_files)
                 monitor_logs(log_files, vcf_files, output_location)
             queue.task_done()
 
@@ -253,8 +251,6 @@ def startup_multiprocessing(
         process.start()
         processes.append(process)
     for files in file_sets:
-        print("Add files to queue:")
-        print(files)
         queue.put(files)
     for core in range(cores):
         queue.put(None)
@@ -345,7 +341,7 @@ def generate_minimal_priority_range():
         minimal priority scores in the exomiser settings, the list is returned.
     """
     minimal_priority_score = []
-    for i in numpy.arange(0.01, 1.0, 0.01):
+    for i in numpy.arange(0.01, 1.0, 0.5):
         minimal_priority_score.append(float("%.2f" % i))
     minimal_priority_score[0] = 0.01
     minimal_priority_score.append(1.0)
@@ -604,10 +600,6 @@ def main():
         user_arguments.cores,
     )
 
-
-def run_exomiser(
-    yaml, vcf, log, temp, config, docker, jar, basename, output_location
-):
 
 if __name__ == "__main__":
     main()