+++ CHANGELOG.md
@@ -59,3 +59,5 @@ Contact information: info@jboom.org.
   to determine the optimal threshold.
 + Add a README for the python scripts.
 + Added a object oriented programming tutorial for python, followed the steps.
++ Convert exomiser script to multiple OOP scripts, including a script to
+  visualise the ROC plot.